+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Ledger/Extended.hs
@@ -29,8 +29,8 @@ module Ouroboros.Consensus.Ledger.Extended (
   , Ticked1 (..)
   ) where
 
-import           Cardano.Ledger.Binary.Plain (Decoder, Encoding,
-                     decodeRecordNamed, encodeListLen)
+import           Codec.CBOR.Decoding (Decoder, decodeListLenOf)
+import           Codec.CBOR.Encoding (Encoding, encodeListLen)
 import           Control.Monad.Except
 import           Data.Coerce
 import           Data.Functor ((<&>))
@@ -211,8 +211,8 @@ decodeExtLedgerState :: (forall s. Decoder s (LedgerState    blk mk))
                      -> (forall s. Decoder s (ExtLedgerState blk mk))
 decodeExtLedgerState decodeLedgerState
                      decodeChainDepState
-                     decodeAnnTip =
-    decodeRecordNamed "ExtLedgerState" (const 2) $ do
+                     decodeAnnTip = do
+      decodeListLenOf 2
       ledgerState <- decodeLedgerState
       headerState <- decodeHeaderState'
       return ExtLedgerState{..}