+++ ouroboros-consensus/src/ouroboros-consensus/Ouroboros/Consensus/Mempool/Update.hs
@@ -7,9 +7,6 @@ module Ouroboros.Consensus.Mempool.Update (
     implAddTx
   , implRemoveTxs
   , implSyncWithLedger
-    -- * Exported for deprecated modules
-  , pureRemoveTxs
-  , pureSyncWithLedger
   ) where
 
 import           Cardano.Slotting.Slot
@@ -24,9 +21,7 @@ import           Ouroboros.Consensus.Block.Abstract (castHash, castPoint,
 import           Ouroboros.Consensus.HeaderValidation
 import           Ouroboros.Consensus.Ledger.Abstract
 import           Ouroboros.Consensus.Ledger.SupportsMempool
-import           Ouroboros.Consensus.Mempool.API hiding (MempoolCapacityBytes,
-                     MempoolCapacityBytesOverride, MempoolSize,
-                     TraceEventMempool, computeMempoolCapacity)
+import           Ouroboros.Consensus.Mempool.API
 import           Ouroboros.Consensus.Mempool.Capacity
 import           Ouroboros.Consensus.Mempool.Impl.Common
 import           Ouroboros.Consensus.Mempool.TxSeq (TxTicket (..))